Move a bunch of GCP functionality into gcpclient.js
