+++ lib/googlechart.js
